backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0

SQLAlchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

pydantic==2.9.2
pydantic-settings==2.6.1

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
